ZONG-110 finished: Introduce Java 8 in all projects

+++ musicxml-in/src/com/xenoage/zong/io/musicxml/in/readers/BarlineReader.java
@@ -57,7 +57,7 @@ public class BarlineReader {
 			}
 			else if (repeat.getDirection() == MxlBackwardForward.Backward) {
 				style = notNull(style, BarlineStyle.LightHeavy);
-				int times = notNull(repeat.getTimes(), 1).intValue();
+				int times = notNull(repeat.getTimes(), 1);
 				barline = barlineBackwardRepeat(style, times);
 			}
 		}
